+++ changelog.md
@@ -1,3 +1,6 @@
+# 0.1.1 - Update: ignore multiple .gitignore files
+- previously wasn't respecting multiple .gitignore files
+
 # 0.1.0 - Add: initial commit
 - creates a codebrev.md file which has a summary of the codebase
 - walks all files and folders for .go and .js files